+++ modules/kbetsy/controllers/front/cron.php
@@ -12,7 +12,10 @@
  * @license   see file: LICENSE.txt
  * @category  PrestaShop Module
  */
-
+//First condition to check if PS Version defined
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
 require_once(_PS_ROOT_DIR_ . '/init.php');
 require_once(_PS_MODULE_DIR_ . 'kbetsy/classes/EtsyModule.php');
 
@@ -28,14 +31,33 @@ class KbetsyCronModuleFrontController extends ModuleFrontController
         @set_time_limit(0);
 
         parent::init();
-        
+
         if (Configuration::get('KBETSY_DEMO')) {
             echo $this->module->l('This is an demo version. Synchronization is not allowed in the demo version.', 'cron');
             die();
         }
 
         $action = Tools::getValue('action');
-        $etsyOAuthAccessToken = Configuration::get('etsy_oauth_access_token');
+        /**
+         * Check if access token is exists or not
+         * @date 09-04-2023
+         * @author Tanisha Gupta
+         */
+        $kb_etsy_token = Configuration::get('kb_etsy_token');
+        $etsyOAuthAccessToken = 1;
+        if (empty($kb_etsy_token)) {
+            $etsyOAuthAccessToken = 0;
+        } else {
+            $token_data_array = json_decode($kb_etsy_token, true);
+            if (isset($token_data_array['access_token'])) {
+                if ($token_data_array['access_token'] == 'null') {
+                    $etsyOAuthAccessToken = 0;
+                }
+            } else {
+                $etsyOAuthAccessToken = 0;
+            }
+        }
+
         $etsySwitchValue = Configuration::get('etsy_switch_value');
 
         if (empty($etsySwitchValue)) {
@@ -63,6 +85,9 @@ class KbetsyCronModuleFrontController extends ModuleFrontController
                         break;
                     case 'syncShopSections':
                         $this->syncShopSections();
+                        break;
+                    case 'syncReturnPolicies':
+                        $this->syncReturnPolicies();
                         break;
                     case 'syncCountriesRegions':
                         $this->syncCountriesRegions();
@@ -94,7 +119,7 @@ class KbetsyCronModuleFrontController extends ModuleFrontController
                 die();
             }
         }
-        echo $this->module->l('Success', 'cron');
+        echo $this->module->l('Sync Process Completed.', 'cron');
         die();
     }
 
@@ -107,7 +132,85 @@ class KbetsyCronModuleFrontController extends ModuleFrontController
             'etsy_api_host' => !Tools::isEmpty(trim(Tools::getValue('etsy_api_host'))) ? Tools::getValue('etsy_api_host') : Configuration::get('etsy_api_host'),
             'etsy_api_version' => !Tools::isEmpty(trim(Tools::getValue('etsy_api_version'))) ? Tools::getValue('etsy_api_version') : Configuration::get('etsy_api_version'),
         );
-        EtsyModule::etsyTestConnection($apiData);
+        /**
+         * Changes done to generate the access token from the etsy and saved into the database.
+         * @date 14-04-2023
+         * @author Tanisha Gupta
+         */
+        $auditLogEntryString = 'Job execution statrted to setup connection with Etsy Marketplace.';
+        $auditMethodName = 'EtsyModule::etsyTestConnection()';
+        EtsyModule::auditLogEntry($auditLogEntryString, $auditMethodName);
+        $etsyRequestURI = '/users/__SELF__';
+        $etsyRequestMethod = 'GET';
+        $etsyQueryString = array();
+        $redirect_url = Context::getContext()->link->getModuleLink('kbetsy', 'cron', array('action' => 'testConnection'));
+        $redirect_url = str_replace("&", "%26", $redirect_url);
+        $config = $apiData['etsy_api_key'];
+        $secure_key = Configuration::get('KBETSY_SECURE_KEY');
+        $code_verifier = $secure_key . $secure_key;
+        $scope = 'address_r address_w billing_r cart_r cart_w email_r feedback_r listings_d listings_r listings_w profile_r profile_w recommend_r recommend_w shops_r shops_w transactions_r transactions_w';
+        $code = '';
+        if (!empty(Tools::getValue('code'))) {
+            $authCode = Tools::getValue('code');
+        }
+        if ((Tools::getValue('cron') == 'token_generate') && empty($authCode)) {
+            $challenge_bytes = hash("sha256", $code_verifier, true);
+            $code_challenge = rtrim(strtr(base64_encode($challenge_bytes), "+/", "-_"), "=");
+            $location = 'https://www.etsy.com/oauth/connect?response_type=code&client_id=' . $config . '&redirect_uri=' . $redirect_url . '&state=' . $secure_key . '&code_challenge=' . $code_challenge . '&code_challenge_method=S256&scope=address_r address_w billing_r cart_r cart_w email_r feedback_r listings_d listings_r listings_w profile_r profile_w recommend_r recommend_w shops_r shops_w transactions_r transactions_w';
+            Tools::redirect($location);
+            die();
+        } elseif (!empty($authCode)) {
+            $auditLogEntryString = 'Job execution completed to setup connection with Etsy Marketplace.';
+            $auditMethodName = 'EtsyModule::etsyTestConnection()';
+            EtsyModule::auditLogEntry($auditLogEntryString, $auditMethodName);
+            $url = 'https://api.etsy.com/v3/public/oauth/token';
+            $headers = [
+                'Content-Type: application/x-www-form-urlencoded'
+            ];
+            $redirect_url = Context::getContext()->link->getModuleLink('kbetsy', 'cron', array('action' => 'testConnection'));
+            $body = http_build_query([
+                'client_id' => $config,
+                'grant_type' => 'authorization_code',
+                'code' => $authCode,
+                'redirect_uri' => $redirect_url,
+                'code_verifier' => $code_verifier
+
+            ]);
+
+            $curl = curl_init();
+            curl_setopt_array($curl, array(
+                CURLOPT_URL => $url,
+                CURLOPT_RETURNTRANSFER => true,
+                CURLOPT_CUSTOMREQUEST => 'POST',
+                CURLOPT_POSTFIELDS => $body,
+                CURLOPT_HTTPHEADER => $headers
+            ));
+
+            $response = curl_exec($curl);
+            $err = curl_error($curl);
+            curl_close($curl);
+
+            if (empty($err)) {
+                $token_data_array = json_decode($response, true);
+                if (isset($token_data_array['access_token'])) {
+                    Configuration::updateGlobalValue('kb_etsy_token', $response);
+                    $token = $token_data_array['access_token'];
+                    $user_id = explode('.', $token);
+                    Configuration::updateGlobalValue('etsy_api_user_id', $user_id[0]);
+                    Tools::redirect(Configuration::get('etsy_redirect_url') . '&etsyConf=1');
+                } else {
+                    Configuration::updateGlobalValue('kb_etsy_token', '');
+                    Tools::redirect(Configuration::get('etsy_redirect_url') . '&etsyError=1');
+                }
+            } else {
+                //If connection failed
+                Configuration::updateGlobalValue('kb_etsy_token', '');
+                Tools::redirect(Configuration::get('etsy_redirect_url') . '&etsyError=1');
+            }
+        } else {
+            echo $this->module->l('Error!!! Please try again.', 'cron');
+            die;
+        }
     }
 
     //To sync shipping templates and shipping templates entries
@@ -125,6 +228,16 @@ class KbetsyCronModuleFrontController extends ModuleFrontController
         SyncShopSection::deleteShopSections();
     }
 
+    /*
+     * Added syncReturnPolicies method to sync return policies from Etsy
+     * @modifier Himanshu Vishwakarma 
+     * @date 15-12-2025
+     */
+    private function syncReturnPolicies()
+    {
+        SyncReturnPolicy::syncEtsyReturnPolicies();
+    }
+
     //To sync countries and regions
     private function syncCountriesRegions()
     {
@@ -140,26 +253,48 @@ class KbetsyCronModuleFrontController extends ModuleFrontController
     /** Populate products on etsy table from PS table based on matching categories of profile */
     private function localSync()
     {
-        EtsyModule::getAllProfileProducts();
+        /**
+         * To add row action for local sync in case if the profile id is present, then sync the products of that profile only
+         * @modifier Pragya Maurya
+         * @date 26-05-2024
+         * PMMay2024 ebay-profile-level-sync
+         */
+        $id_profile = false;
+        if (!empty(Tools::getValue('profile_id'))) {
+            $id_profile = Tools::getValue('profile_id');
+        }
+        EtsyModule::getAllProfileProducts($id_profile);
     }
 
     //To sync the new products  on etsy
     private function syncProductsListing()
     {
-
         $language = Configuration::get('etsy_default_lang') != '' ? Configuration::get('etsy_default_lang') : Context::getContext()->language->id;
         $listingArray = array();
-
         $id_product = false;
         if (!empty(Tools::getValue('id_product'))) {
             $id_product = Tools::getValue('id_product');
         }
 
-        /* Call delete product funcation before syncing the product */
-        $this->syncDeleteProductsListing($id_product);
+        /**
+         * Changes added by pragya maurya for localsync on profile level
+         * Etsy-enhancement-profile-level
+         * @modifier pragya maurya
+         * @date 04-06-2024
+         */
+        $id_profile = false;
+        if (!empty(Tools::getValue('profile_id'))) {
+            $id_profile = Tools::getValue('profile_id');
+        }
 
-        $products = EtsyModule::getProductsToListOnEtsy($this->limit, $id_product);
-        
+        /**
+         * to add the id_profile as well in the syncDeleteProductsListing function so that the products of that profile only get deleted
+         * Etsy-enhancement-profile-level
+         * @modifier pragya maurya
+         * @date 04-06-2024
+         */
+        $this->syncDeleteProductsListing($id_product, $id_profile);
+        $products = EtsyModule::getProductsToListOnEtsy($this->limit, $id_product, $id_profile);
         if (isset($products) && count($products) > 0) {
             foreach ($products as $product) {
                 $product_data = EtsyModule::prepareArrayToListingOnEtsy($product, $language);
@@ -173,16 +308,28 @@ class KbetsyCronModuleFrontController extends ModuleFrontController
         }
     }
 
-    //To delete the products on etsy which has been marked as deleted on the Prestashop module
-    private function syncDeleteProductsListing($kbproductid)
+
+    /**
+     * Modified changes to delete the products of that profile only
+     * Etsy-enhancement-profile-level
+     * @modifier pragya maurya
+     * @date 04-06-2024
+     */
+    private function syncDeleteProductsListing($kbproductid, $kbprofileid)
     {
-        $productsList = EtsyModule::getProductsToDeleteOnEtsy($kbproductid);
+        $productsList = EtsyModule::getProductsToDeleteOnEtsy($kbproductid, $kbprofileid);
+
         if (!empty($productsList)) {
             foreach ($productsList as $product) {
                 if (!empty($product['listing_id'])) {
-                    EtsyModule::deleteItemsFromEtsy($product);
+                    //changes by gopi
+                    //non need to delete profile with 0 profile id
+                    if ($product['id_etsy_profiles'] != 0) {
+                        EtsyModule::deleteItemsFromEtsy($product);
+                    }
+                    //change by gopi end here
                 } else {
-                    DB::getInstance()->execute("DELETE FROM " . _DB_PREFIX_ . "etsy_products_list WHERE id_product = '" . (int) $product['id_product'] . "'");
+                    Db::getInstance()->execute("DELETE FROM " . _DB_PREFIX_ . "etsy_products_list WHERE id_product = '" . (int) $product['id_product'] . "'");
                 }
             }
         }
@@ -192,30 +339,8 @@ class KbetsyCronModuleFrontController extends ModuleFrontController
     //To sync update products listing status
     private function syncProductsListingStatus()
     {
-        $productsList = EtsyModule::getProductsListedOnEtsy();
-        if ($productsList > 0) {
-            EtsyModule::syncItemListingStatus();
-        }
-
-        /* Below code is commented as logic of status sync is changed. Fetching all the items from the etsy & updating the status instead of checking the stauts of Individual Item */
-        /**
-          $listingArray = array();
-          if (isset($productsList) && count($productsList) > 0) {
-          foreach ($producsyncProductsListingStatustsList as $productsList) {
-          $listingArray[] = array(
-          'listing_id' => $productsList['listing_id'],
-          );
-          }
-          }
-
-          if (isset($listingArray) && count($listingArray) > 0) {
-          if (EtsyModule::etsyGetListings($listingArray)) {
-          return true;
-          }
-          } else {
-          return true;
-          }
-         */
+        EtsyModule::getProductsListedOnEtsy();
+        EtsyModule::syncItemListingStatus();
     }
 
     //To sync orders from etsy to PS
@@ -236,7 +361,7 @@ class KbetsyCronModuleFrontController extends ModuleFrontController
 
         $language = Configuration::get('etsy_default_lang') != '' ? Configuration::get('etsy_default_lang') : Context::getContext()->language->id;
         $lang_data = new Language($language);
-        
+
         $importFile = _PS_MODULE_DIR_ . 'kbetsy/data/categories_' . Tools::strtolower($lang_data->iso_code) . '.sql';
         if (!file_exists($importFile)) {
             $importFile = _PS_MODULE_DIR_ . 'kbetsy/data/categories_en.sql';
@@ -285,11 +410,11 @@ class KbetsyCronModuleFrontController extends ModuleFrontController
                 $final_name = $category['name'];
             }
             Db::getInstance()->execute("INSERT INTO " . _DB_PREFIX_ . "etsy_categories SET "
-                    . "category_code = '" . (int) $category['id'] . "',"
-                    . "tag = '" . pSQL($category['path']) . "',"
-                    . "category_name = '" . pSQL($final_name) . "',"
-                    . "property_set = '',"
-                    . "parent_id = '" . (int) $parent_id . "'");
+                . "category_code = '" . (int) $category['id'] . "',"
+                . "tag = '" . pSQL($category['path']) . "',"
+                . "category_name = '" . pSQL($final_name) . "',"
+                . "property_set = '',"
+                . "parent_id = '" . (int) $parent_id . "'");
             $category_inserted_id = Db::getInstance()->Insert_ID();
 
             if ($category_inserted_id) {
@@ -297,8 +422,8 @@ class KbetsyCronModuleFrontController extends ModuleFrontController
                     $this->importCategory($category_inserted_id, $final_name, true, $category['children']);
                 } else {
                     Db::getInstance()->execute("UPDATE " . _DB_PREFIX_ . "etsy_categories SET "
-                            . "last_level = 1 "
-                            . "WHERE id_etsy_categories = '" . (int) $category_inserted_id . "'");
+                        . "last_level = 1 "
+                        . "WHERE id_etsy_categories = '" . (int) $category_inserted_id . "'");
                 }
             }
         }
